api/api/requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
react==4.3.0
reportlab==3.6.8